requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.200
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
docling>=1.0.0
requests>=2.28.0
numpy>=1.24.0
langchain-community>=0.3.24
accelerate>=1.7.0
